Mutation API: extract MutationRecorder off the client
Concept: snapshot/publish is not the Mutation Client's concern. Move
it to a MutationRecorder that lives on DashboardScene.

Wins:
- one recorder per dashboard regardless of caller (UI client, future
  MCP gateway, plugin host)
- a clean seam for future audit/replay tooling to read from
- a per-dashboard ownership boundary natural for multiplayer

DashboardMutationClient now owns dispatch (lookup, permission, lock,
validation, per-domain queue). It calls recorder.record(opts, mutate)
which frames the handler call with snapshot/restore + event publish.

Behavior identical to the pre-extraction client. Conflict detection,
coalescing, addedObject/removedObject diff all move into the recorder
as the natural home.
